requirements.txt
numpy
pillow
qrcode
moviepy
imageio
imageio-ffmpeg